Regularized polynomial regression
Regularized polynomial regression with linear and random sampling.

We take an example of a single-variable, non-polynomial function approximation and show that in the presence of strong regularization, both sampling methods seem to work reasonably well, even when the sample density is small. However, without regularization i.e. simple polynomial regression tends to favor the uniform (linearly spaced) sampling method on the average. We further show, that this difference goes away when the sample density is made larger or the measurement/observation noise is made small.